Create user feature and add user related files into it
