+++ src/features/user/Login.js
@@ -1,10 +1,12 @@
 import React from "react";
+// import { useHistory } from "react-router";
 import { Link } from "react-router-dom";
 import FederatedLoginButton from "./FederatedLoginButton";
 import styles from "./Login.module.css";
-import logo from './logo.svg';
+import logo from '../../logo.svg';
 
-function Register() {
+
+function Login() {
   return (
     <section className={styles.root}>
       <header className={styles.logoHeader}>
@@ -12,15 +14,11 @@ function Register() {
       </header>
       <section class={styles.mainContent}>
         <article class={styles.innerContent}>
-          <h1 className={styles.brandlabel}>Sign up for your account</h1>
+          <h1 className={styles.brandlabel}>Log in to Trello</h1>
           <form>
             <input type="email" name="email" placeholder="Enter email" required />
-            <input type="text" name="full_name" placeholder="Enter full name" required />
             <input type="password" name="password" placeholder="Enter password" required />
-
-            <p className={styles.descriptors}>By signing up, you confirm that you've read and accepted our Terms of Service and Privacy Policy.</p>
-
-            <button type="submit">Sign up</button>
+            <button type="submit">Log in</button>
           </form>
           <p className={styles.descriptors}>OR</p>
           <FederatedLoginButton cssId={'google-auth'} >
@@ -30,7 +28,8 @@ function Register() {
             Continue with Slack
           </FederatedLoginButton>
           <ul className={styles.bottomLink}>
-            <li><Link to="/login">Already have an account? Log In</Link></li>
+            <li><Link to="/">Can't log in?</Link></li>
+            <li><Link to="/signup">Sign up for an account</Link></li>
           </ul>
         </article>
       </section>
@@ -38,4 +37,4 @@ function Register() {
   );
 }
 
-export default Register;
+export default Login;
