+++ src/features/user/Register.js
@@ -1,12 +1,10 @@
 import React from "react";
-// import { useHistory } from "react-router";
 import { Link } from "react-router-dom";
 import FederatedLoginButton from "./FederatedLoginButton";
 import styles from "./Login.module.css";
-import logo from './logo.svg';
+import logo from '../../logo.svg';
 
-
-function Login() {
+function Register() {
   return (
     <section className={styles.root}>
       <header className={styles.logoHeader}>
@@ -14,11 +12,15 @@ function Login() {
       </header>
       <section class={styles.mainContent}>
         <article class={styles.innerContent}>
-          <h1 className={styles.brandlabel}>Log in to Trello</h1>
+          <h1 className={styles.brandlabel}>Sign up for your account</h1>
           <form>
             <input type="email" name="email" placeholder="Enter email" required />
+            <input type="text" name="full_name" placeholder="Enter full name" required />
             <input type="password" name="password" placeholder="Enter password" required />
-            <button type="submit">Log in</button>
+
+            <p className={styles.descriptors}>By signing up, you confirm that you've read and accepted our Terms of Service and Privacy Policy.</p>
+
+            <button type="submit">Sign up</button>
           </form>
           <p className={styles.descriptors}>OR</p>
           <FederatedLoginButton cssId={'google-auth'} >
@@ -28,8 +30,7 @@ function Login() {
             Continue with Slack
           </FederatedLoginButton>
           <ul className={styles.bottomLink}>
-            <li><Link to="/">Can't log in?</Link></li>
-            <li><Link to="/signup">Sign up for an account</Link></li>
+            <li><Link to="/login">Already have an account? Log In</Link></li>
           </ul>
         </article>
       </section>
@@ -37,4 +38,4 @@ function Login() {
   );
 }
 
-export default Login;
+export default Register;
